Updated bets to be compatible with back end gamemanager. App is now ready for full testing on a local/azure server.

+++ frontend/AndroidCasinoUser/app/src/main/java/com/davidzhang/androidcasinouser/BaccaratActivity.java
@@ -251,7 +251,7 @@ public class BaccaratActivity extends AppCompatActivity {
                             double earnings = 0;
                             try {
                                 JSONObject gameResult = gameResults.getJSONObject("gameResult");
-                                earnings = gameResult.getInt(userName);
+                                earnings = gameResult.getDouble(userName);
                             } catch (Exception e) {
                                 e.printStackTrace();
                             }
